split up dht
